requirements.txt
fastapi==0.136.0
uvicorn==0.45.0
httpx==0.28.1
pydantic-settings==2.14.0
python-telegram-bot==22.7
pyupbit==0.2.34
pyngrok==8.0.0
structlog==25.5.0
